+++ Plugins/ImageToClipboard/ImageToClipboard.plugin.js
@@ -1,12 +1,12 @@
 //META{"name":"ImageToClipboard"}*//
 
-/* global PluginUtilities:false, ReactUtilities:false */
+/* global PluginUtilities:false, ReactUtilities:false, PluginContextMenu:false, DiscordModules:false */
 
 class ImageToClipboard {
 	getName() { return "ImageToClipboard"; }
 	getShortName() { return "i2c"; }
 	getDescription() { return "Copies images (png/jpg) directly to clipboard. Support Server: bit.ly/ZeresServer"; }
-	getVersion() { return "0.2.11"; }
+	getVersion() { return "0.2.12"; }
 	getAuthor() { return "Zerebos"; }
 
 	constructor() {
@@ -17,8 +17,8 @@ class ImageToClipboard {
 		this.path = require("path");
 		this.fileSystem = require("fs");
 		this.process = require("process");
-		this.link = '<a target="_blank" rel="noreferrer" class="downloadLink-wANcd8 size14-1wjlWP weightMedium-13x9Y8">${modalLabel}</a>';
-		this.contextItem = '<div class="itemGroup-oViAgA i2c-group"><div class="item-1XYaYf i2c-item"><span>${contextMenuLabel}</span><div class="hint"></div></div></div>';
+		this.link = '<a target="_blank" rel="noreferrer" class="downloadLink-2oSgiF size14-3iUx6q weightMedium-2iZe9B">${modalLabel}</a>';
+		// this.contextItem = '<div class="itemGroup-oViAgA i2c-group"><div class="item-1XYaYf i2c-item"><span>${contextMenuLabel}</span><div class="hint"></div></div></div>';
 	}
 	
 	load() {}
@@ -105,18 +105,18 @@ class ImageToClipboard {
 		if (!e.addedNodes.length || !(e.addedNodes[0] instanceof Element) || !this.initialized) return;
 		var elem = $(e.addedNodes[0]);
 
-		if (elem.hasClass("modal-2LIEKY") && elem.find(".inner-1_1f7b .imageWrapper-38T7d9").length) {
+		if (elem.hasClass("modal-1UGdnR") && elem.find(".inner-1JeGVc .imageWrapper-2p5ogY").length) {
 			var linkElement = $(PluginUtilities.formatString(this.link, this.strings));
-			var openElement = $('.inner-1_1f7b>div>div>a');
+			var openElement = $('.inner-1JeGVc>div>div>a');
 			var originalLink = openElement.attr("href").replace(/:large$/, '');
 			var imageLink = originalLink.split('?')[0];
 			if (imageLink.endsWith('.png') || imageLink.endsWith('.jpg') || imageLink.endsWith('.jpeg')) {
-				openElement.after($('<span class="downloadLink-wANcd8 size14-1wjlWP weightMedium-13x9Y8"> | </span>'), linkElement);
+				openElement.after($('<span class="downloadLink-2oSgiF size14-3iUx6q weightMedium-2iZe9B"> | </span>'), linkElement);
 				linkElement.on("click", () => { this.copyToClipboard(originalLink); });
 			}
 		}
 
-		if (elem.hasClass("contextMenu-uoJTbz")) {
+		if (elem.hasClass(DiscordModules.ContextMenuClasses.contextMenu)) {
 			this.bindMenu(elem[0]);
 		}