+++ Plugins/ServerSearch/CHANGELOG.md
@@ -1,5 +1,10 @@
 # ServerSearch Changelog
 
+### 0.0.3
+
+ - Fix for discord update
+ - Use internal classes
+
 ### 0.0.2
 
  - Update for discord changes